ref(onboarding): Load getting started doc by clicking on the platform
